Add files via upload
Power consumption in different zones in a city area can be affected by weather parameters, the landscape and the type of occupancy. In this project, we will analyze the data of power consumption in zone 1 in a city named Wellington in New Zealand. We will find best fitted data model to predict the power consumptions in the zone 1.